Created another file for edit mode handling function
